refactor: moved data parsers to own classes
