+++ src/Pretzel.Logic/Templating/Context/DataParsing/IDataParser.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Linq;
+
+namespace Pretzel.Logic.Templating.Context.DataParsing
+{
+    internal interface IDataParser
+    {
+        string Extension { get; }
+
+        bool CanParse(string folder, string method);
+
+        object Parse(string folder, string method);
+    }
+}
